refact/front/UserDetailsForm: rename from CreateAccountOne.jsx to UserDetailsForm.jsx
