+++ frontend/src/pages/register/user/UserDetailsForm.jsx
